Struct deduped_dashmap::DedupedMap

pub struct DedupedMap<K, M, V> where
    K: Eq + Hash,
    M: Debug,
    V: Debug,  {
    pointers: DashMap<K, MapPointer<M>>,
    storage: DashMap<u64, MapValue<V>>,
}

A hashmap that assumes a large number of keys will map to a relatively smaller number of values.

If a value is stored in the map that is already stored by another key, the refcount of the value will be incremented instead of duplicating it.

Uses DashMap internally

Fields
pointers: DashMap<K, MapPointer<M>>

First layer of the map. A mapping from incoming requests to a u64 pointer into the cache storage.

storage: DashMap<u64, MapValue<V>>

Second layer of the map. The items stored in the cache, keyed by their hash.

Implementations

impl<K, M, V> DedupedMap<K, M, V> where
K: Eq + Hash + Debug,
M: Debug + Clone,
V: Hash + Debug + Clone,

pub fn new() -> Self

Create an empty map.

pub fn insert(&self, key: K, meta: M, value: V)

Insert value into the map at key with metadata meta.

If value is already in the map under a different key, its refcount will be incremented instead of storing another copy of value. The metadata data meta will be attached to this specific key, and not refcounted.

pub fn remove(&self, key: K)

Remove the item associated with a key from the map.

The metadata associated with key will always be removed. The reference count of the storage item it points to will be decremented. If no more keys refer to the storage item, it will also be removed.

pub fn get(&self, key: &K) -> Option<(M, V)>

Get cloned copies of the metadata and value associated with key.

pub fn len_storage(&self) -> usize

Fetches the total number of storage items in the map.

This will be at most self.len_pointers().

pub fn len_pointers(&self) -> usize

Fetches the total number of pointers stored in the map.

This will be at least self.len_storage().

pub fn retain<F>(&self, pred: F) where
F: FnMut(&K, &M, &V) -> ControlFlow<bool>,

Retain elements based on the result of a predicate.

If the predicate function returns:

ControlFlow::Continue(true): The item will be retained and the iteration will continue.

ControlFlow::Continue(false): The item will be removed and the iteration will continue.

ControlFlow::Break(()): This item, and all further items in the map, will be retained, and the predicate won’t be called again.

pub fn contains_key(&self, key: &K) -> bool

Checks if the map contains a specific key.